1. Add tests for humanizing negative values. 2. Move all speech tests validated through assertions (rather than vocalizations) from `SpeechTests.cs` to `SpeechUnitTests.cs`.
