+++ Tests/SpeechTests.cs
@@ -134,12 +134,6 @@ namespace SpeechTests
         }
 
         [TestMethod, TestCategory("Speech")]
-        public void TestTranslationVesper()
-        {
-            Assert.AreEqual(Translations.StarSystem("VESPER-M4"), "Vesper M 4");
-        }
-
-        [TestMethod, TestCategory("Speech")]
         public void TestPowerplay()
         {
             var ship = ShipDefinitions.FromEliteID(128049363);
@@ -233,157 +227,6 @@ namespace SpeechTests
             SpeechService.Instance.Say(ShipDefinitions.FromEliteID(128049339), "Anaconda golf foxtrot lima one niner six eight returning from orbit.", 3, null, true);
         }
 
-        [TestMethod, TestCategory("Speech")]
-        public void TestPathingString1()
-        {
-            string pathingString = @"There are [4;5] lights";
-            List<string> pathingOptions = new List<string>() {
-                "There are 4 lights"
-                , "There are 5 lights"
-            };
-
-            HashSet<string> pathingResults = new HashSet<string>();
-            for (int i = 0; i < 1000; i++)
-            {
-                string pathedString = VoiceAttackPlugin.SpeechFromScript(pathingString);
-                pathingResults.Add(pathedString);
-            }
-
-            Assert.IsTrue(pathingResults.SetEquals(new HashSet<string>(pathingOptions)));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestPathingString2()
-        {
-            string pathingString = @"There are [4;5;] lights";
-            List<string> pathingOptions = new List<string>() {
-                "There are 4 lights"
-                , "There are 5 lights"
-                , "There are  lights"
-            };
-
-            HashSet<string> pathingResults = new HashSet<string>();
-            for (int i = 0; i < 1000; i++)
-            {
-                string pathedString = VoiceAttackPlugin.SpeechFromScript(pathingString);
-                pathingResults.Add(pathedString);
-            }
-
-            Assert.IsTrue(pathingResults.SetEquals(new HashSet<string>(pathingOptions)));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestPathingString3()
-        {
-            string pathingString = @"There [are;might be;could be] [4;5;] lights;It's dark in here;";
-            List<string> pathingOptions = new List<string>() {
-                "There are 4 lights"
-                , "There are 5 lights"
-                , "There are  lights"
-                ,"There might be 4 lights"
-                , "There might be 5 lights"
-                , "There might be  lights"
-                ,"There could be 4 lights"
-                , "There could be 5 lights"
-                , "There could be  lights"
-                , "It's dark in here"
-                , ""
-            };
-
-            HashSet<string> pathingResults = new HashSet<string>();
-            for (int i = 0; i < 1000; i++)
-            {
-                string pathedString = VoiceAttackPlugin.SpeechFromScript(pathingString);
-                pathingResults.Add(pathedString);
-            }
-
-            Assert.IsTrue(pathingResults.SetEquals(new HashSet<string>(pathingOptions)));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestPathingString4()
-        {
-            string pathingString = @";;;;;;Seven;;;";
-            List<string> pathingOptions = new List<string>() {
-                ""
-                , "Seven"
-            };
-
-            int sevenCount = 0;
-            for (int i = 0; i < 10000; i++)
-            {
-                string pathedString = VoiceAttackPlugin.SpeechFromScript(pathingString);
-                if (pathedString == "Seven")
-                {
-                    sevenCount++;
-                }
-            }
-
-            Assert.IsTrue(sevenCount > 750);
-            Assert.IsTrue(sevenCount < 1500);
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestPathingString5()
-        {
-            string pathingString = @"You leave me [no choice].";
-            List<string> pathingOptions = new List<string>() {
-                "You leave me no choice."
-            };
-
-            HashSet<string> pathingResults = new HashSet<string>();
-            for (int i = 0; i < 1000; i++)
-            {
-                string pathedString = VoiceAttackPlugin.SpeechFromScript(pathingString);
-                pathingResults.Add(pathedString);
-            }
-
-            Assert.IsTrue(pathingResults.SetEquals(new HashSet<string>(pathingOptions)));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestPathingString6()
-        {
-            string pathingString = @"[There can be only one.]";
-            List<string> pathingOptions = new List<string>() {
-                "There can be only one."
-            };
-
-            HashSet<string> pathingResults = new HashSet<string>();
-            for (int i = 0; i < 1000; i++)
-            {
-                string pathedString = VoiceAttackPlugin.SpeechFromScript(pathingString);
-                pathingResults.Add(pathedString);
-            }
-
-            Assert.IsTrue(pathingResults.SetEquals(new HashSet<string>(pathingOptions)));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestPathingString7()
-        {
-            string pathingString = @"[{TXT:Ship model} {TXT:Ship callsign (spoken)};This is {TXT:Ship model} {TXT:Ship callsign (spoken)}] [requesting docking permission;requesting docking clearance;requesting permission to dock;requesting clearance to dock].";
-            List<string> pathingOptions = new List<string>() {
-                "{TXT:Ship model} {TXT:Ship callsign (spoken)} requesting docking permission."
-                ,"This is {TXT:Ship model} {TXT:Ship callsign (spoken)} requesting docking permission."
-                ,"{TXT:Ship model} {TXT:Ship callsign (spoken)} requesting docking clearance."
-                ,"This is {TXT:Ship model} {TXT:Ship callsign (spoken)} requesting docking clearance."
-                ,"{TXT:Ship model} {TXT:Ship callsign (spoken)} requesting clearance to dock."
-                ,"This is {TXT:Ship model} {TXT:Ship callsign (spoken)} requesting clearance to dock."
-                ,"{TXT:Ship model} {TXT:Ship callsign (spoken)} requesting permission to dock."
-                ,"This is {TXT:Ship model} {TXT:Ship callsign (spoken)} requesting permission to dock."
-            };
-
-            HashSet<string> pathingResults = new HashSet<string>();
-            for (int i = 0; i < 1000; i++)
-            {
-                string pathedString = VoiceAttackPlugin.SpeechFromScript(pathingString);
-                pathingResults.Add(pathedString);
-            }
-
-            Assert.IsTrue(pathingResults.SetEquals(new HashSet<string>(pathingOptions)));
-        }
-
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2202:Do not dispose objects multiple times")] // this usage is perfecrtly correct    
         [TestMethod, TestCategory("Speech")]
         public void TestSpeech()
@@ -435,73 +278,6 @@ namespace SpeechTests
         {
             Logging.Verbose = true;
             SpeechService.Instance.Speak("You are  docked at Jameson Memorial  in the <phoneme alphabet=\"ipa\" ph=\"ʃɪnˈrɑːrtə\">Shinrarta</phoneme> <phoneme alphabet=\"ipa\" ph=\"ˈdezɦrə\">Dezhra</phoneme> system.", null, 50, 1, 30, 40, 0, true);
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSectorTranslations()
-        {
-            Assert.AreEqual("Swoiwns N Y dash B a 95 dash 0", Translations.StarSystem("Swoiwns NY-B a95-0"));
-            Assert.AreEqual("P P M 5 2 8 7", Translations.StarSystem("PPM 5287"));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize1()
-        {
-            Assert.AreEqual("on the way to 12 and a half thousand", Translations.Humanize(12345));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize2()
-        {
-            Assert.AreEqual(null, Translations.Humanize(null));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize3()
-        {
-            Assert.AreEqual("zero", Translations.Humanize(0));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize4()
-        {
-            Assert.AreEqual("0.16", Translations.Humanize(0.15555555M));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize5()
-        {
-            Assert.AreEqual("0.016", Translations.Humanize(0.015555555M));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize6()
-        {
-            Assert.AreEqual("0.0016", Translations.Humanize(0.0015555555M));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize7()
-        {
-            Assert.AreEqual("51 million", Translations.Humanize(51000000));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize8()
-        {
-            Assert.AreEqual("just over 51 million", Translations.Humanize(51000001));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize9()
-        {
-            Assert.AreEqual("10 thousand", Translations.Humanize(10000));
-        }
-
-        [TestMethod, TestCategory("Speech")]
-        public void TestSpeechHumanize10()
-        {
-            Assert.AreEqual("100 thousand", Translations.Humanize(100000));
         }
 
         [TestMethod, TestCategory("Speech")]
